requirements.txt
aioredis==1.2.0
asgiref==3.1.2
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.3.3
Automat==0.7.0
certifi==2019.3.9
channels==2.2.0
channels-redis==2.4.0
chardet==3.0.4
constantly==15.1.0
daphne==2.3.0
dj-database-url==0.5.0
Django==2.2
django-cors-headers==2.5.2
django-mysql==3.0.0.post1
django-nose==1.4.6
djangorestframework==3.9.2
djangorestframework-camel-case==1.0.3
factory-boy==2.11.1
Faker==1.0.5
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
msgpack==0.6.1
msgpack-python==0.5.6
mysqlclient==1.4.2.post1
nose==1.3.7
oauthlib==3.0.1
PyHamcrest==1.9.0
python-dateutil==2.8.0
pytz==2019.1
redis==2.10.6
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
Twisted==19.2.0
txaio==18.8.1
urllib3==1.24.2
zope.interface==4.6.0
